Rename xtest to xrecord.
