+++ x11-dl/src/lib.rs
@@ -27,4 +27,4 @@ pub mod xlib;
 pub mod xmu;
 pub mod xrender;
 pub mod xt;
-pub mod xtest;
+pub mod xrecord;
