--- x11-dl/src/xtest.rs
@@ -1,1 +1,0 @@
-../../src/xtest.rs
