+++ x11/build.rs
@@ -14,5 +14,5 @@ fn main () {
   if cfg!(feature="xmu") { let _ = pkg_config::find_library("xmu"); }
   if cfg!(feature="xrender") { let _ = pkg_config::find_library("xrender"); }
   if cfg!(feature="xt") { let _ = pkg_config::find_library("xt"); }
-  if cfg!(feature="xtest") { let _ = pkg_config::find_library("xtst"); }
+  if cfg!(feature="xrecord") { let _ = pkg_config::find_library("xtst"); }
 }
