+++ x11/src/lib.rs
@@ -24,4 +24,4 @@ pub mod xlib;
 pub mod xmu;
 pub mod xrender;
 pub mod xt;
-pub mod xtest;
+pub mod xrecord;
